[core][android] Add enum event support to OnStartObserving abd OnStopObserving
